requirements.txt
future~=0.18.2
pillow~=8.0.1
imageio~=2.9.0
numpy~=1.19.2
matplotlib~=3.3.2
seaborn~=0.11.0
pandas~=1.1.3
networkx~=2.5